refactor: ui package 수정
