+++ app/src/main/java/com/naccoro/wask/ui/calendar/CalendarAdapter.java
@@ -1,4 +1,4 @@
-package com.naccoro.wask.calendar;
+package com.naccoro.wask.ui.calendar;
 
 import android.content.Context;
 import android.util.Log;
@@ -13,7 +13,7 @@ import com.naccoro.wask.preferences.SettingPreferenceManager;
 import com.naccoro.wask.replacement.repository.ReplacementHistoryRepository;
 import com.naccoro.wask.utils.AlarmUtil;
 import com.naccoro.wask.utils.DateUtils;
-import com.naccoro.wask.calendar.CalendarActivity.Date;
+import com.naccoro.wask.ui.calendar.CalendarActivity.Date;
 
 import java.util.ArrayList;
 import java.util.Calendar;
