+++ app/src/main/java/com/naccoro/wask/ui/calendar/CalendarContract.java
@@ -1,7 +1,7 @@
-package com.naccoro.wask.calendar;
+package com.naccoro.wask.ui.calendar;
 
 import java.util.ArrayList;
-import com.naccoro.wask.calendar.CalendarActivity.Date;
+import com.naccoro.wask.ui.calendar.CalendarActivity.Date;
 
 public interface CalendarContract {
     interface View {
