+++ app/src/main/java/com/naccoro/wask/ui/calendar/CalendarItem.java
@@ -1,4 +1,4 @@
-package com.naccoro.wask.calendar;
+package com.naccoro.wask.ui.calendar;
 
 import java.util.GregorianCalendar;
