+++ app/src/main/java/com/naccoro/wask/ui/calendar/CalendarModel.java
@@ -1,4 +1,4 @@
-package com.naccoro.wask.calendar;
+package com.naccoro.wask.ui.calendar;
 
 import android.util.Log;
 
@@ -7,7 +7,7 @@ import java.util.ArrayList;
 import java.util.GregorianCalendar;
 import java.util.List;
 
-import com.naccoro.wask.calendar.CalendarActivity.Date;
+import com.naccoro.wask.ui.calendar.CalendarActivity.Date;
 import com.naccoro.wask.replacement.repository.ReplacementHistoryRepository;
 import com.naccoro.wask.utils.DateUtils;
