+++ app/src/main/res/layout/activity_calendar.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/waskBlue"
-    tools:context=".calendar.CalendarActivity">
+    tools:context=".ui.calendar.CalendarActivity">
 
     <com.naccoro.wask.customview.WaskToolbar
         android:id="@+id/wasktoolbar_calendar"
