+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/waskBlue"
-    tools:context=".main.MainActivity">
+    tools:context=".ui.main.MainActivity">
 
     <com.naccoro.wask.customview.WaskToolbar
         android:id="@+id/wasktoolbar_main"
